Added flags to clothes in preparation to the genitals system
hide_boobs = 0 //Used to see if it hides breasts  or not.
hide_genitals = 0 //Used to see if it hides genitals or not.

hope I didn't missed any, but there is a posibility I did.

new clothes that aren't derivate of the big clothes families should get
those vars on 1, and thus hide naugthy parts